Moved method and tests
